reorder classes
