+++ Rdmp.Core/Datasets/HDRDatasetPatch.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Rdmp.Core.Datasets.HDRItems;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
