+++ Rdmp.Core/Datasets/HDRDatasetProvider.cs
@@ -1,10 +1,16 @@
-﻿using Amazon.Runtime.Internal.Endpoints.StandardLibrary;
+﻿// Copyright (c) The University of Dundee 2025-2025
+// This file is part of the Research Data Management Platform (RDMP).
+// RDMP is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+// RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+// You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
+using Amazon.Runtime.Internal.Endpoints.StandardLibrary;
 using Newtonsoft.Json;
 using NPOI.OpenXmlFormats.Wordprocessing;
 using Org.BouncyCastle.Asn1.Cms;
 using Rdmp.Core.CommandExecution;
 using Rdmp.Core.Curation.Data;
 using Rdmp.Core.Curation.Data.Datasets;
+using Rdmp.Core.Datasets.HDRItems;
 using System;
 using System.Collections.Generic;
 using System.Data;
@@ -18,7 +24,9 @@ using System.Threading.Tasks;
 
 namespace Rdmp.Core.Datasets
 {
-
+    /// <summary>
+    /// Provider for connecting to the HDR Gateway
+    /// </summary>
     public class HDRDatasetProvider : PluginDatasetProvider
     {
         private HttpClient _client;
