+++ Rdmp.Core/Datasets/HDRItems/Accessibility.cs
@@ -1,0 +1,13 @@
+﻿
+namespace Rdmp.Core.Datasets.HDRItems
+{
+ /// <summary>
+ /// 
+ /// </summary>
+    public class Accessibility
+    {
+        public Access access { get; set; }
+        public Usage usage { get; set; }
+        public FormatAndStandards formatAndStandards { get; set; }
+    }
+}
