+++ Rdmp.Core/Datasets/HDRItems/DataCustodian.cs
@@ -1,0 +1,15 @@
+﻿namespace Rdmp.Core.Datasets.HDRItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class DataCustodian
+    {
+        public string name { get; set; }
+        public string identifier { get; set; }
+        public string contactPoint { get; set; }
+        public object logo { get; set; }
+        public object description { get; set; }
+        public object memberOf { get; set; }
+    }
+}
