+++ Rdmp.Core/Datasets/HDRItems/Documentation.cs
@@ -1,0 +1,12 @@
+﻿namespace Rdmp.Core.Datasets.HDRItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class Documentation
+    {
+        public string description { get; set; }
+        public object associatedMedia { get; set; }
+        public string inPipeline { get; set; }
+    }
+}
