+++ Rdmp.Core/Datasets/HDRItems/HDRSummary.cs
@@ -1,4 +1,9 @@
-﻿using System;
+﻿// Copyright (c) The University of Dundee 2025-2025
+// This file is part of the Research Data Management Platform (RDMP).
+// RDMP is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+// RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+// You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,9 +11,13 @@ using System.Threading.Tasks;
 
 namespace Rdmp.Core.Datasets.HDRItems
 {
+    /// <summary>
+    /// HDR simple summary object
+    /// </summary>
     public class HDRSummary
     {
-        public HDRSummary(string title) {
+        public HDRSummary(string title)
+        {
             Title = title;
         }
