+++ Rdmp.Core/Datasets/HDRItems/Provenance.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.HDRItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class Provenance
+    {
+        public Origin origin { get; set; }
+        public Temporal temporal { get; set; }
+    }
+}
