+++ Rdmp.Core/Datasets/HDRItems/Revision.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.HDRItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class Revision
+    {
+        public string version { get; set; }
+        public string url { get; set; }
+    }
+}
