+++ Rdmp.Core/Datasets/HDRItems/SpatialCoverage.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.HDRItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class SpatialCoverage
+    {
+        public int id { get; set; }
+        public DateTime created_at { get; set; }
+        public DateTime updated_at { get; set; }
+        public string region { get; set; }
+        public bool enabled { get; set; }
+        public List<int> dataset_version_ids { get; set; }
+    }
+}
