+++ Rdmp.Core/Datasets/JiraItems/DefaultType.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.JiraItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class DefaultType
+    {
+        public int id { get; set; }
+        public string name { get; set; }
+    }
+}
