+++ Rdmp.Core/Datasets/JiraItems/ExtendedInfo.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.JiraItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class ExtendedInfo
+    {
+        public bool openIssuesExists { get; set; }
+        public bool attachmentsExists { get; set; }
+    }
+}
