+++ Rdmp.Core/Datasets/JiraItems/Icon.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.JiraItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class Icon
+    {
+        public string id { get; set; }
+        public string name { get; set; }
+        public string url16 { get; set; }
+        public string url48 { get; set; }
+    }
+}
