+++ Rdmp.Core/Datasets/JiraItems/Links.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.JiraItems
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class Links
+    {
+        public string self { get; set; }
+    }
+}
