+++ Rdmp.Core/Datasets/JiraItems/ObjectType.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rdmp.Core.Datasets.JiraItems
+{
+
+
+    /// <summary>
+    /// 
+    /// </summary>
+    public class ObjectType
+    {
+        public string workspaceId { get; set; }
+        public string globalId { get; set; }
+        public string id { get; set; }
+        public string name { get; set; }
+        public int type { get; set; }
+        public string description { get; set; }
+        public Icon icon { get; set; }
+        public int position { get; set; }
+        public DateTime created { get; set; }
+        public DateTime updated { get; set; }
+        public int objectCount { get; set; }
+        public string objectSchemaId { get; set; }
+        public bool inherited { get; set; }
+        public bool abstractObjectType { get; set; }
+        public bool parentObjectTypeInherited { get; set; }
+    }
+}
