+++ Rdmp.UI.Tests/DesignPatternTests/ClassFileEvaluation/ExplicitDatabaseNameChecker.cs
@@ -29,7 +29,9 @@ public class ExplicitDatabaseNameChecker
             "PluginPackagerProgramOptions.cs", //allowed because it's a suggestion to the user about command line arguments
             "DocumentationCrossExaminationTest.cs", //allowed because its basically a list of comments that are allowed despite not appearing in the codebase
             "ResearchDataManagementPlatformOptions.cs", //allowed because it's an Example
-            "AWSS3BucketReleaseDestination.cs" //allowed as it uses it as a temp file identifier
+            "AWSS3BucketReleaseDestination.cs", //allowed as it uses it as a temp file identifier
+            "JiraDatasetProvider.cs" //jira user the prohibited string
+
         };
